requirements.txt
fastapi
uvicorn
huggingface-hub
requests
python-dotenv
gunicorn